move `webpack` configs out of `scripts`
they’re not scripts, they’re tooling config, like `playwright` or `cypress`
