+++ dotcom-rendering/webpack/.swcrc.json
