+++ dotcom-rendering/webpack/browser-targets.js
