+++ dotcom-rendering/webpack/svg.cjs
