+++ dotcom-rendering/webpack/webpack.config.server.js
